+++ app/src/main/java/de/tomcory/heimdall/scanner/traffic/connection/appLayer/DnsConnection.kt
@@ -3,7 +3,7 @@ package de.tomcory.heimdall.scanner.traffic.connection.appLayer
 import de.tomcory.heimdall.scanner.traffic.connection.encryptionLayer.EncryptionLayerConnection
 import timber.log.Timber
 
-class DnsConnection(id: Int, encryptionLayer: EncryptionLayerConnection) : AppLayerConnection(id, encryptionLayer) {
+class DnsConnection(id: Long, encryptionLayer: EncryptionLayerConnection) : AppLayerConnection(id, encryptionLayer) {
 
     init {
         Timber.w("%s Creating DNS connection", id)
